requirements.txt
python-telegram-bot==20.7
pytest
pytest-cov
python-dotenv
pylint
mypy
flake8
isort
